+++ ibm/provider/provider.go
@@ -418,6 +418,7 @@ func Provider() *schema.Provider {
 			"ibm_is_region":                      vpc.DataSourceIBMISRegion(),
 			"ibm_is_regions":                     vpc.DataSourceIBMISRegions(),
 			"ibm_is_ssh_key":                     vpc.DataSourceIBMISSSHKey(),
+			"ibm_is_ssh_keys":                    vpc.DataSourceIBMIsSshKeys(),
 			"ibm_is_subnet":                      vpc.DataSourceIBMISSubnet(),
 			"ibm_is_subnets":                     vpc.DataSourceIBMISSubnets(),
 			"ibm_is_subnet_reserved_ip":          vpc.DataSourceIBMISReservedIP(),
@@ -431,6 +432,7 @@ func Provider() *schema.Provider {
 			"ibm_is_snapshot":                    vpc.DataSourceSnapshot(),
 			"ibm_is_snapshots":                   vpc.DataSourceSnapshots(),
 			"ibm_is_volume":                      vpc.DataSourceIBMISVolume(),
+			"ibm_is_volumes":                     vpc.DataSourceIBMIsVolumes(),
 			"ibm_is_volume_profile":              vpc.DataSourceIBMISVolumeProfile(),
 			"ibm_is_volume_profiles":             vpc.DataSourceIBMISVolumeProfiles(),
 			"ibm_is_vpc":                         vpc.DataSourceIBMISVPC(),
